Move TextButton factory to TextButton class file
